Partial revert of the debian patch, enable multiple auth types
Properly  check the requested auth type and return DECLINED if this
request is not meant to be handled by this module.

The issue is documented here:
http://sourceforge.net/p/authmemcookie/bugs/5/

The Auth_memcookie module breaks basic authentication, which, at least
on my system, comes after this module in the request sequence. If
Auth_memcookie receives a request with an AuthType other than
"Cookie," it returns HTTP_UNAUTHORIZED. I believe it should return
DECLINED so that the request can be passed to the next module.

Thank you Steve Gaarder